update detail map control panel
